fix(admin): move admin role validation to token utils
